SimpleLink WIFI 9.22.00.12 wifi examples update This commit updates wifi examples to 9.22.00.12 WIFI SDK version

+++ examples/rtos/LP_EM_CC35X1/demos/network_terminal/adaptation/uart_term.c
@@ -320,7 +320,7 @@ void InitTerm(void)
 
 void UART_writePolling(uint8_t *buffer, uint32_t len)
 {
-    uint8_t uartIter;
+    uint32_t uartIter;
 
     for (uartIter=0; uartIter<len; uartIter++)
     {
